+++ src/ShaMap/ShaMap.ts
@@ -1,0 +1,3 @@
+import ShaMapInner from './ShaMapInner'
+
+export default class ShaMap extends ShaMapInner {}
