+++ src/ShaMap/index.ts
@@ -1,0 +1,3 @@
+export { default as ShaMapNode } from './ShaMapNode'
+export { default as ShaMapLeaf } from './ShaMapLeaf'
+export { default } from './ShaMap'
