+++ src/hashes.ts
@@ -2,7 +2,7 @@ import { Buffer } from 'buffer/'
 import * as createHash from 'create-hash'
 
 import HashPrefix from './hash-prefixes'
-import { BytesList } from './serdes/binary-serializer'
+import BytesList from './serdes/BytesList'
 import Hash256 from './types/hash-256'
 
 /**
@@ -11,7 +11,7 @@ import Hash256 from './types/hash-256'
  * @extends BytesList So SerializedTypes can write bytes to a Sha512Half
  */
 class Sha512Half extends BytesList {
-  private hash: createHash = createHash('sha512')
+  private hash = createHash('sha512')
 
   /**
    * Construct a new Sha512Hash and write bytes this.hash.
@@ -40,7 +40,7 @@ class Sha512Half extends BytesList {
    * @returns Half of a SHA512 hash.
    */
   finish256(): Buffer {
-    const bytes: Buffer = this.hash.digest()
+    const bytes = Buffer.from(this.hash.digest())
     return bytes.slice(0, 32)
   }
