+++ src/ledger-hashes.ts
@@ -5,10 +5,10 @@ import * as bigInt from 'big-integer'
 import { BinarySerializer, serializeObject } from './binary'
 import HashPrefix from './hash-prefixes'
 import { Sha512Half } from './hashes'
-import BinaryParser from './serdes/binary-parser'
-import { ShaMap, ShaMapNode, ShaMapLeaf } from './shamap'
+import BinaryParser from './serdes/BinaryParser'
+import ShaMap, { ShaMapNode, ShaMapLeaf } from './ShaMap'
 import Hash256 from './types/hash-256'
-import { JsonObject } from './types/serialized-type'
+import { JsonObject } from './types/SerializedType'
 import STObject from './types/st-object'
 import UInt32 from './types/uint-32'
 import UInt64 from './types/uint-64'
