+++ src/serdes/BinarySerializer.ts
@@ -4,57 +4,14 @@ import * as assert from 'assert'
 import { Buffer } from 'buffer/'
 
 import { FieldInstance } from '../enums'
-import { SerializedType } from '../types/serialized-type'
+import SerializedType from '../types/SerializedType'
 
-/**
- * Bytes list is a collection of buffer objects.
- */
-class BytesList {
-  private bytesArray: Buffer[] = []
-
-  /**
-   * Get the total number of bytes in the BytesList.
-   *
-   * @returns The number of bytes.
-   */
-  public getLength(): number {
-    return Buffer.concat(this.bytesArray).byteLength
-  }
-
-  /**
-   * Put bytes in the BytesList.
-   *
-   * @param bytesArg - A Buffer.
-   * @returns This BytesList.
-   */
-  public put(bytesArg: Buffer): BytesList {
-    const bytes = Buffer.from(bytesArg) // Temporary, to catch instances of Uint8Array being passed in
-    this.bytesArray.push(bytes)
-    return this
-  }
-
-  /**
-   * Write this BytesList to the back of another bytes list.
-   *
-   *  @param list - The BytesList to write to.
-   */
-  public toBytesSink(list: BytesList): void {
-    list.put(this.toBytes())
-  }
-
-  public toBytes(): Buffer {
-    return Buffer.concat(this.bytesArray)
-  }
-
-  toHex(): string {
-    return this.toBytes().toString('hex').toUpperCase()
-  }
-}
+import BytesList from './BytesList'
 
 /**
  * BinarySerializer is used to write fields and values to buffers.
  */
-class BinarySerializer {
+export default class BinarySerializer {
   private sink: BytesList = new BytesList()
 
   constructor(sink: BytesList) {
@@ -153,5 +110,3 @@ class BinarySerializer {
     this.writeBytesList(bytes)
   }
 }
-
-export { BytesList, BinarySerializer }
