+++ test/hash.test.js
@@ -1,6 +1,6 @@
 const { Buffer } = require('buffer/')
 
-const { Hash160, Hash256, AccountID, Currency } = require('../dist/types').default
+const { Hash160, Hash256, AccountID, Currency } = require('../dist/types')
 
 describe('Hash160', function () {
   test('has a static width member', function () {
