+++ test/uint.test.js
@@ -1,5 +1,5 @@
 const { encode } = require('../dist')
-const { UInt8, UInt64 } = require('../dist/types').default
+const { UInt8, UInt64 } = require('../dist/types')
 
 const binary =
   '11007222000300003700000000000000003800000000000000006280000000000000000000000000000000000000005553440000000000000000000000000000000000000000000000000166D5438D7EA4C680000000000000000000000000005553440000000000AE123A8556F3CF91154711376AFB0F894F832B3D67D5438D7EA4C680000000000000000000000000005553440000000000F51DFC2A09D62CBBA1DFBDD4691DAC96AD98B90F'
