+++ tsconfig.json
@@ -1,27 +1,27 @@
 {
-    "compilerOptions": {
-        "target": "es5",
-        "lib": [
-        "es2017"
-        ],
-        "outDir": "dist",
-        "rootDir": "src",
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "noUnusedLocals": true,
-        "noUnusedParameters": true,
-        "resolveJsonModule": true,
-        "noImplicitThis": false,
-        "noImplicitAny": false,
-        "removeComments": false,
-        "preserveConstEnums": false,
-        "suppressImplicitAnyIndexErrors": true,
-        "sourceMap": true,
-        "skipLibCheck": true,
-        "declaration": true,
-        "strict": true
-    },
-    "include": [
-        "src/**/*.ts"
-    ]
+  "compilerOptions": {
+    "target": "es5",
+    "lib": [
+      "es2017"
+    ],
+    "outDir": "dist",
+    "rootDir": "src",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "resolveJsonModule": true,
+    "noImplicitThis": false,
+    "noImplicitAny": false,
+    "removeComments": false,
+    "preserveConstEnums": false,
+    "suppressImplicitAnyIndexErrors": true,
+    "sourceMap": true,
+    "skipLibCheck": true,
+    "declaration": true,
+    "strict": true
+  },
+  "include": [
+    "src/**/*.ts"
+  ]
 }
